refactor(ui): split out Dialog and DialogConfirm into separate components
